requirements.txt
# веб-фреймворк
Django==4.2.17
# REST API
djangorestframework>=3.11
# JWT-авторизация
djangorestframework-simplejwt==4.8.0
# работа с .env
django-environ==0.12.0
python-dotenv==1.0.1
# Celery + результаты (храним в Redis)
celery==5.5.3
django-celery-results==2.6.0
redis==6.2.0
# асинхронные HTTP-запросы из задач
aiohttp==3.11.13
# синхронные HTTP-запросы (тестовый клиент)
requests==2.20.0
# WSGI-сервер
gunicorn==23.0.0
# (опционально) если у вас есть static/media файлы
Pillow==11.1.0
drf-spectacular>=0.27.2
openai>=1.40.0
httpx>=0.27.0
tenacity>=8.3.0
pybreaker>=1.0.2
psycopg2-binary>=2.9.9
django-pgvector>=0.2.5
